TFP-4873 overstyrings AP for ny fakta om uttak
